Change utils of Mantra.js to a global file and added according tests
