+++ tests/unit/Mantra.spec.js
@@ -2,7 +2,7 @@ import 'jest-extended';
 import Vuex from 'vuex';
 import { createLocalVue, shallowMount } from '@vue/test-utils';
 import { MantraPlugin } from '../../lib/MantraPlugin';
-import Mantra, { componentsRegistration, install, isEmptyObject, isObject, isOptionsValid, isTruthy, isValidSchema, isValidStore, registerStoreModule } from '../../Mantra';
+import Mantra, { componentsRegistration, install, isOptionsValid, isValidSchema, isValidStore, registerStoreModule } from '../../Mantra';
 
 describe('Mantra.js module', () => {
 	it('should export a function', () => {
@@ -11,68 +11,6 @@ describe('Mantra.js module', () => {
 
 	it('should export a function call `install`', () => {
 		expect(Mantra.name).toEqual('install');
-	});
-
-	describe('isTruthy', () => {
-
-		it('should be a function', () => {
-			expect(isTruthy).toBeInstanceOf(Function);
-		});
-
-		it('should return `true` for truthy values', () => {
-			expect(isTruthy('hola')).toBeTrue();
-			expect(isTruthy(3)).toBeTrue();
-			expect(isTruthy(true)).toBeTrue();
-			expect(isTruthy({})).toBeTrue();
-			expect(isTruthy([])).toBeTrue();
-			expect(isTruthy(new Function())).toBeTrue();
-		});
-
-		it('should return `false` for falsy values', () => {
-			expect(isTruthy('')).toBeFalse();
-			expect(isTruthy(false)).toBeFalse();
-			expect(isTruthy(0)).toBeFalse();
-			expect(isTruthy(null)).toBeFalse();
-			expect(isTruthy(undefined)).toBeFalse();
-		});
-	});
-
-	describe('isObject', () => {
-
-		it('should be a function', () => {
-			expect(isObject).toBeInstanceOf(Function);
-		});
-
-		it('should return `true` for object values', () => {
-			expect(isObject({ pepe: 5, rodrigo: 6 })).toBeTrue();
-			expect(isObject(new Object(null))).toBeTrue();
-		});
-
-		it('should return `false` for non-object values', () => {
-			expect(isObject([])).toBeFalse();
-			expect(isObject(2)).toBeFalse();
-			expect(isObject('hola')).toBeFalse();
-			expect(isObject(true)).toBeFalse();
-			expect(isObject(new Function())).toBeFalse();
-			expect(isObject(new Map())).toBeFalse();
-			expect(isObject(new Set())).toBeFalse();
-			expect(isObject(new Promise(() => true))).toBeFalse();
-		});
-	});
-
-	describe('isEmptyObject', () => {
-
-		it('should be a function', () => {
-			expect(isEmptyObject).toBeInstanceOf(Function);
-		});
-
-		it('should return `true` for an object with no properties', () => {
-			expect(isEmptyObject({})).toBeTrue();
-		});
-
-		it('should return `false` for an object with properties', () => {
-			expect(isEmptyObject({ pepe: 'hola' })).toBeFalse();
-		});
 	});
 
 	describe('isValidSchema', () => {
